refactor: 관련된 도메인 별로 패키지 변경
