+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -1,6 +1,5 @@
-package kitchenpos.menuproduct.domain;
+package kitchenpos.menu.domain;
 
-import kitchenpos.menu.domain.Menu;
 import kitchenpos.product.domain.Product;
 
 import javax.persistence.Entity;
