+++ src/main/java/kitchenpos/ordertable/domain/OrderTable.java
@@ -1,7 +1,5 @@
 package kitchenpos.ordertable.domain;
 
-import kitchenpos.tablegroup.domain.TableGroup;
-
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
