+++ src/main/java/kitchenpos/ordertable/repository/TableGroupRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.tablegroup.repository;
+package kitchenpos.ordertable.repository;
 
-import kitchenpos.tablegroup.domain.TableGroup;
+import kitchenpos.ordertable.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
