+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -2,12 +2,12 @@ package kitchenpos.application;
 
 import kitchenpos.order.repository.OrderRepository;
 import kitchenpos.ordertable.repository.OrderTableRepository;
-import kitchenpos.tablegroup.repository.TableGroupRepository;
+import kitchenpos.ordertable.repository.TableGroupRepository;
 import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.tablegroup.domain.TableGroup;
-import kitchenpos.tablegroup.application.TableGroupService;
-import kitchenpos.tablegroup.application.dto.OrderTableId;
-import kitchenpos.tablegroup.application.dto.TableGroupCreateRequest;
+import kitchenpos.ordertable.domain.TableGroup;
+import kitchenpos.ordertable.application.TableGroupService;
+import kitchenpos.ordertable.application.dto.OrderTableId;
+import kitchenpos.ordertable.application.dto.TableGroupCreateRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
