+++ src/test/java/kitchenpos/domain/MenuProductsTest.java
@@ -1,7 +1,7 @@
 package kitchenpos.domain;
 
-import kitchenpos.menuproduct.domain.MenuProduct;
-import kitchenpos.menuproduct.domain.MenuProducts;
+import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProducts;
 import kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
